live data is working
