+++ src/server/db/schema.ts
@@ -1,3 +1,4 @@
+import "server-only";
 import {
   int,
   bigint,
